fix: Fixed

+++ src/Modules/Tours/Explorer.Tours.API/Dtos/KeyPointDto.cs
@@ -5,5 +5,5 @@
     public string Name { get; set; }
     public string? Description { get; set; }
     public string? Image { get; set; } 
-    public int TourId { get; set; }
+    public long TourId { get; set; }
 }
